requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
coloredlogs==15.0.1
fastapi==0.110.0
filelock==3.13.1
flatbuffers==24.3.6
fsspec==2024.2.0
h11==0.14.0
humanfriendly==10.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
onnxruntime==1.17.1
packaging==23.2
pillow==10.2.0
protobuf==4.25.3
pydantic==2.6.3
pydantic_core==2.16.3
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.10.0
uvicorn==0.27.1
